seikei

+++ common/src/main/java/dev/felnull/miningunderworld/block/MiningTntBlock.java
@@ -41,7 +41,7 @@ public class MiningTntBlock extends TntBlock {
     @Override
     public void wasExploded(Level level, BlockPos blockPos, Explosion explosion) {
         if (!level.isClientSide) {
-            PrimedMiningTnt primedTnt = new PrimedMiningTnt(level, (double) blockPos.getX() + 0.5, (double) blockPos.getY(), (double) blockPos.getZ() + 0.5, explosion.getIndirectSourceEntity());
+            PrimedMiningTnt primedTnt = new PrimedMiningTnt(level, (double) blockPos.getX() + 0.5, blockPos.getY(), (double) blockPos.getZ() + 0.5, explosion.getIndirectSourceEntity());
             int i = primedTnt.getFuse();
             primedTnt.setFuse((short) (level.random.nextInt(i / 4) + i / 8));
             level.addFreshEntity(primedTnt);
